CAMEL-4796: Removed not needed type casts. Thanks to Babak for the patch.
git-svn-id: https://svn.apache.org/repos/asf/camel/trunk@1227658 13f79535-47bb-0310-9956-ffa450edef68

+++ camel-core/src/test/java/org/apache/camel/issues/SplitterParallelIssueTest.java
@@ -61,7 +61,7 @@ public class SplitterParallelIssueTest extends ContextTestSupport {
                             @Override
                             public void process(Exchange exchange) throws Exception {
                                 int num = exchange.getIn().getBody(int.class);
-                                final long sleep = (long) (num * delay);
+                                final long sleep = num * delay;
                                 log.info("Sleep for " + sleep + "ms");
                                 Thread.sleep(sleep);
                             }
